authenticate: fix bugzilla.redhat.com issues (4.4.x)
There are two problems in the authenticate() subroutine.

The first is that git-send-bugzilla sends a GET request, even though it
should send a POST. This is remedied by replacing "mech->submit" with the
more explicit "mech->submit_form", where the login button is explicitly
selected for clicking.

The second problem is documented here:

https://jira.almworks.com/browse/DZO-1189

Basically, the POST request for logging in must contain the valid
Bugzilla_login_token value that the initial GET returned.

Unfortunately, the initial GET doesn't return such a Bugzilla_login_token
at all, because Bugzilla 4.4.x (erroneously) provides that only if the
initial GET contains Bugzilla_login_request_cookie. Since we don't do that
on the initial get, we receive no Bugzilla_login_token, and then the POST
request is rejected.

This is fixed by sending the initial GET twice -- the first GET will makes
sure we set Bugzilla_login_request_cookie on the second, and the second
(now correct) GET makes sure that Bugzilla_login_token is set in the
response. We then reflect Bugzilla_login_token in the login POST request.

Signed-off-by: Laszlo Ersek <lersek@redhat.com>